requirements.txt
django>=3.0.4,<4.0
django-environ>=0.4.5,<1.0
django-bootstrap-static>=4.2.1,<5.0
plexapi>=3.3.0,<4.0
jinja2>=2.11.1,<3.0
requests
